[feature/loan-delay-management] Adicionar gerenciamento de multas de empréstimos
Alteracoes:
- backend/controllers/BookController.js
  - Validar se autoresId e genresId não são arrays vazios.
- backend/controllers/LibraryController.js
  - Remover validação de campos duplicada.
- backend/controllers/LoanController.js
  - Implementar métodos para obter e atualizar multas de empréstimos.
- backend/controllers/UserController.js
  - Adicionar validação para alteração de cargo de usuário.
- backend/helpers/cronTabs/verifyOverdueCron.js
  - Criar cron para verificar empréstimos vencidos.
- backend/models/Loan/LoanDelay.js
  - Adicionar novo modelo para gerenciar multas de empréstimos.
- backend/routes/LoanRoutes.js
  - Adicionar rotas para gerenciamento de multas de empréstimos.
- backend/services/BookService.js
  - Adicionar validações para exclusão de autores, gêneros e livros.
- backend/services/LoanService.js
  - Implementar lógica para verificar e atualizar empréstimos vencidos e suas multas.
- frontend/src/hooks/useLoanDelay.jsx
  - Criar hook para gerenciar chamadas de API relacionadas a multas de empréstimos.
- frontend/src/routes/Loan/LoanDelay/EditLoanDelay.jsx
  - Criar componente para editar multas de empréstimos.
- frontend/src/routes/Loan/Loan/LoanItem.jsx
  - Adicionar botão para visualizar multas associadas a empréstimos.
- frontend/src/routes/Loan/Loan/LoanItem.module.css
  - Estilizar botão de visualização de multas.
- frontend/src/routes/Loan/Loan/ListLoan.jsx
  - Alterar título de "Exemplares" para "Empréstimos".

Notas:
- Implementação do gerenciamento de multas de empréstimos, incluindo verificação automática de vencimentos.
- Sem observacoes adicionais.

+++ backend/services/BookService.js
@@ -186,6 +186,15 @@ module.exports = class BookService {
             return existingPublisher
         }
 
+        const bookWithPublisher = await Book.findOne({ publisherId: existingPublisher.publisher._id })
+        if (bookWithPublisher) {
+            return {
+                valid: false,
+                message: 'Não é possível excluir a editora pois existem livros cadastrados',
+                err: 'publisher-has-books'
+            }
+        }
+
         try {
             await Publisher.findByIdAndDelete(existingPublisher.publisher._id)
             return {
@@ -357,6 +366,15 @@ module.exports = class BookService {
         const existingAuthor = await this.existingAuthor(authorData)
         if (existingAuthor && !existingAuthor.valid) {
             return existingAuthor
+        }
+
+        const existingBooks = await Book.find({ authorsId: existingAuthor.author._id })
+        if (existingBooks.length > 0) {
+            return {
+                valid: false,
+                message: 'Não é possível excluir o autor pois existem livros cadastrados',
+                err: 'author-has-books'
+            }
         }
 
         try {
@@ -605,6 +623,15 @@ module.exports = class BookService {
             return {
                 valid: false,
                 message: 'Não é possível excluir um gênero com filhos'
+            }
+        }
+
+        const existingBooks = await Book.find({ genresId: existingGenre.genre._id })
+        if (existingBooks.length > 0) {
+            return {
+                valid: false,
+                message: 'Não é possível excluir o gênero pois existem livros cadastrados',
+                err: 'genre-has-books'
             }
         }
 
@@ -889,6 +916,15 @@ module.exports = class BookService {
             return existingBook
         }
 
+        const existingBookCopies = await BookCopy.find({ bookId: existingBook.book._id })
+        if (existingBookCopies.length > 0) {
+            return {
+                valid: false,
+                message: 'Não é possível excluir o livro pois existem exemplares cadastrados',
+                err: 'book-has-copies'
+            }
+        }
+
         try {
             await Book.findByIdAndDelete(existingBook.book._id)
             return {
@@ -964,7 +1000,7 @@ module.exports = class BookService {
         try {
             const activeLoan = await Loan.findOne({
                 copyId: bookCopyData.id,
-                status: { $in: ['active', 'overdue'] }
+                status: { $in: ['active', 'overdue', 'lost'] },
             })
             if (activeLoan) {
                 return {
@@ -1215,6 +1251,11 @@ module.exports = class BookService {
         const existingBookCopy = await this.existingBookCopy(bookCopyData)
         if (existingBookCopy && !existingBookCopy.valid) {
             return existingBookCopy
+        }
+
+        const loan = await this.checkIfLoanIsActive(bookCopyData)
+        if (loan && !loan.valid) {
+            return loan
         }
 
         try {
